backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pymongo==4.6.1
PyPDF2==3.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.3
python-docx==1.1.0

# Gemini SDK
google-generativeai==0.5.4


# Libraries for OpenAI API

# fastapi==0.109.0
# uvicorn==0.27.0
# pymongo==4.6.1
# openai==1.12.0
# PyPDF2==3.0.1
# python-multipart==0.0.6
# python-dotenv==1.0.0
# pydantic==2.5.3